// dashboard/src/utils/hasura-api/generated/schemas/createLogicalModelSelectPermissionStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateLogicalModelSelectPermissionArgs } from './createLogicalModelSelectPermissionArgs';
import type { CreateLogicalModelSelectPermissionStepType } from './createLogicalModelSelectPermissionStepType';

export interface CreateLogicalModelSelectPermissionStep {
  type: CreateLogicalModelSelectPermissionStepType;
  args: CreateLogicalModelSelectPermissionArgs;
}